+++ template/trout/config/initializers/will_paginate.rb
@@ -1,4 +1,0 @@
-WillPaginate::ViewHelpers.pagination_options.merge!(
-  :previous_label => '« Vorige Seite',
-  :next_label     => 'Nächste Seite »'
-)